# docs/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Site Availability Monitor API
  version: 1.0.0-design
  summary: Private management and public status APIs
  description: >-
    Canonical Phase 4 design contract. This file describes the intended v1 API;
    it does not claim that the routes are implemented yet. Error semantics are
    defined in API_ERRORS.md and realtime semantics in REALTIME_CONTRACT.md.
servers:
  - url: http://localhost:13000
    description: Local development
...
security:
  - cookieAuth: []
paths:
...
  /api/v1/notification-recipients:
    get:
      tags: [Notifications]
      operationId: listNotificationRecipients
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Notification recipients
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationRecipientPage'
    post:
      tags: [Notifications]
      operationId: createNotificationRecipient
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailRequest'
      responses:
        '201':
          description: Recipient created; verification scheduled
          headers:
            Location:
              $ref: '#/components/headers/Location'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationRecipient'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '422':
          $ref: '#/components/responses/ValidationProblem'
  /api/v1/notification-recipients/{recipient_id}:
    delete:
      tags: [Notifications]
      operationId: deleteNotificationRecipient
      parameters:
        - in: path
          name: recipient_id
          required: true
          schema:
            $ref: '#/components/schemas/Uuid'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '204':
          description: Recipient disabled
        '404':
          $ref: '#/components/responses/NotFoundProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
  /api/v1/notification-recipients/{recipient_id}/verification:
    post:
      tags: [Notifications]
      operationId: resendNotificationRecipientVerification
      parameters:
        - in: path
          name: recipient_id
          required: true
          schema:
            $ref: '#/components/schemas/Uuid'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          $ref: '#/components/responses/Accepted'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '429':
          $ref: '#/components/responses/RateLimitProblem'
  /api/v1/notification-recipients/{recipient_id}/test-email:
    post:
      tags: [Notifications]
      operationId: sendNotificationRecipientTestEmail
      parameters:
        - in: path
          name: recipient_id
          required: true
          schema:
            $ref: '#/components/schemas/Uuid'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          $ref: '#/components/responses/Accepted'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '429':
          $ref: '#/components/responses/RateLimitProblem'
  /api/v1/notification-recipient-verifications/confirm:
    post:
      tags: [Notifications]
      operationId: confirmNotificationRecipientVerification
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '204':
          description: Recipient verified
        '422':
          $ref: '#/components/responses/ValidationProblem'
  /api/v1/notification-policies/default:
    get:
      tags: [Notifications]
      operationId: getDefaultNotificationPolicy
      responses:
        '200':
          description: Default notification policy
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPolicy'
    put:
      tags: [Notifications]
      operationId: replaceDefaultNotificationPolicy
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DefaultNotificationPolicyWrite'
      responses:
        '200':
          description: Policy replaced
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPolicy'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '422':
          $ref: '#/components/responses/ValidationProblem'
  /api/v1/groups/{group_id}/notification-policy:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    get:
      tags: [Notifications]
      operationId: getGroupNotificationPolicy
      responses:
        '200':
          description: Effective group notification policy
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPolicy'
    put:
      tags: [Notifications]
      operationId: replaceGroupNotificationPolicy
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationPolicyWrite'
      responses:
        '200':
          description: Group policy replaced
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPolicy'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
        '422':
          $ref: '#/components/responses/ValidationProblem'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: site_monitor_session
  parameters:
    CsrfToken:
      in: header
      name: X-CSRF-Token
      required: true
      schema:
        type: string
        minLength: 20
        maxLength: 512
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[\x21-\x7E]+$'
    IfMatch:
      in: header
      name: If-Match
      required: true
      schema:
        type: string
        pattern: '^"rv-[0-9]+"$'
...
    Cursor:
      in: query
      name: cursor
      required: false
      schema:
        type: string
        maxLength: 1024
    Limit:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
    GroupId:
      in: path
      name: group_id
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
...
  headers:
    ETag:
      description: Strong resource or projection version tag
      schema:
        type: string
    Location:
      description: Canonical resource URI
      schema:
        type: string
        format: uri-reference
...
  responses:
    Accepted:
      description: Accepted without revealing account existence
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AcceptedResponse'
...
    NotFoundProblem:
      description: Resource absent, deleted, or owned by another user
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ConflictProblem:
      description: Domain, idempotency, or resource conflict
      headers:
        Retry-After:
          description: Present for idempotency_in_progress
          schema:
            type: integer
            minimum: 0
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    PreconditionFailed:
      description: If-Match does not match current resource version
      headers:
        ETag:
          $ref: '#/components/headers/ETag'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationProblem:
      description: Request validation or configuration failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    PreconditionRequired:
      description: If-Match header is required
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    RateLimitProblem:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 0
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    Uuid:
      type: string
      format: uuid
    Instant:
      type: string
      format: date-time
    ResourceVersion:
      type: string
      pattern: '^[0-9]+$'
...
    AcceptedResponse:
      type: object
      additionalProperties: false
      required: [accepted]
      properties:
        accepted:
          type: boolean
          const: true
    ValidationIssue:
      type: object
      additionalProperties: false
      required: [pointer, code, message]
      properties:
        pointer:
          type: string
        code:
          type: string
          pattern: '^[a-z][a-z0-9_]*$'
        message:
          type: string
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, instance, code, request_id, retryable]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
          pattern: '^[a-z][a-z0-9_]*$'
        request_id:
          type: string
          format: uuid
        retryable:
          type: boolean
        retry_after_seconds:
          type: integer
          minimum: 0
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
    PageMeta:
      type: object
      additionalProperties: false
      required: [next_cursor, has_more]
      properties:
        next_cursor:
          type: [string, 'null']
        has_more:
          type: boolean
...
    EmailRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
    TokenRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token:
          type: string
          minLength: 20
          maxLength: 512
...
    NotificationRecipient:
      type: object
      additionalProperties: false
      required: [id, email, verification_state, resource_version, created_at]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        email:
          type: string
          format: email
        verification_state:
          type: string
          enum: [PENDING, VERIFIED, DISABLED]
        resource_version:
          $ref: '#/components/schemas/ResourceVersion'
        created_at:
          $ref: '#/components/schemas/Instant'
    NotificationRecipientPage:
      type: object
      additionalProperties: false
      required: [data, page]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/NotificationRecipient'
        page:
          $ref: '#/components/schemas/PageMeta'
    NotificationPolicyWrite:
      type: object
      additionalProperties: false
      required: [mode, notify_down, notify_recovery, recipient_ids]
      properties:
        mode:
          type: string
          enum: [INHERIT, ACTIVE, DISABLED]
        notify_down:
          type: [boolean, 'null']
        notify_recovery:
          type: [boolean, 'null']
        recipient_ids:
          type: array
          uniqueItems: true
          maxItems: 100
          items:
            $ref: '#/components/schemas/Uuid'
      allOf:
        - if:
            properties:
              mode:
                const: ACTIVE
          then:
            properties:
              notify_down:
                type: boolean
              notify_recovery:
                type: boolean
              recipient_ids:
                type: array
                minItems: 1
          else:
            properties:
              notify_down:
                type: 'null'
              notify_recovery:
                type: 'null'
              recipient_ids:
                type: array
                maxItems: 0
    DefaultNotificationPolicyWrite:
      type: object
      additionalProperties: false
      required: [mode, notify_down, notify_recovery, recipient_ids]
      properties:
        mode:
          type: string
          enum: [ACTIVE, DISABLED]
        notify_down:
          type: [boolean, 'null']
        notify_recovery:
          type: [boolean, 'null']
        recipient_ids:
          type: array
          uniqueItems: true
          maxItems: 100
          items:
            $ref: '#/components/schemas/Uuid'
      allOf:
        - if:
            properties:
              mode:
                const: ACTIVE
          then:
            properties:
              notify_down:
                type: boolean
              notify_recovery:
                type: boolean
              recipient_ids:
                type: array
                minItems: 1
          else:
            properties:
              notify_down:
                type: 'null'
              notify_recovery:
                type: 'null'
              recipient_ids:
                type: array
                maxItems: 0
    NotificationPolicy:
      type: object
      additionalProperties: false
      required:
        [
          id,
          scope_type,
          scope_id,
          mode,
          notify_down,
          notify_recovery,
          recipient_ids,
          effective_recipient_ids,
          effective_policy_id,
          effective_mode,
          effective_notify_down,
          effective_notify_recovery,
          effective_policy_version,
          resource_version,
        ]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        scope_type:
          type: string
          enum: [DEFAULT, GROUP]
        scope_id:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        mode:
          type: string
          enum: [INHERIT, ACTIVE, DISABLED]
        notify_down:
          type: [boolean, 'null']
        notify_recovery:
          type: [boolean, 'null']
        recipient_ids:
          type: array
          items:
            $ref: '#/components/schemas/Uuid'
        effective_recipient_ids:
          type: array
          items:
            $ref: '#/components/schemas/Uuid'
        effective_policy_id:
          $ref: '#/components/schemas/Uuid'
        effective_mode:
          type: string
          enum: [ACTIVE, DISABLED]
        effective_notify_down:
          type: [boolean, 'null']
        effective_notify_recovery:
          type: [boolean, 'null']
        effective_policy_version:
          $ref: '#/components/schemas/ResourceVersion'
        resource_version:
          $ref: '#/components/schemas/ResourceVersion'
...